+++ src/App.js
@@ -2,10 +2,6 @@
 import React from "react";
 import Style from "./App.module.css";
 import RouterComponent from "./component/pages/RouterComponent/RouterComponent";
-import Filter from "./component/Filter/Filter";
-import InformationBoxLayout from "./layouts/InformationBoxLayout/InformationBoxLayout";
-import Header from './component/Header/Header';
-
 
 export default function App() {
   // axios
@@ -17,10 +13,7 @@ export default function App() {
   // 	});
   return (
     <div>
-	  <Header />
       <RouterComponent />
-      <InformationBoxLayout />
-	  <Filter />
     </div>
   );
 }
