+++ src/App.module.css
@@ -1,9 +1,3 @@
 body {
 	overflow: hidden;
 }
-
-.test {
-	background-color: var(--color-green);
-	font-family: 'Franklin Gothic Heavy';
-	font-size: 26px;
-}
