+++ src/component/Filter/Filter.module.css
@@ -2,7 +2,6 @@
   display: flex;
   justify-content: space-around;
   align-items: center;
-  padding-left: 50px;
 }
 
 /* First I'm hiding the default buttons */
@@ -29,7 +28,7 @@
   top: 0;
   width: 30px;
   height: 30px;
-  border: 1px solid #ffffff;
+  border: 1px solid #58595b;
   border-radius: 100%;
 }
