+++ src/component/InformationBoxLayout/InformationBoxLayout.module.css
@@ -4,8 +4,8 @@
   background: rgba(255, 255, 255, 0.5);
   border-radius: 45px;
   position: absolute;
-  top: 10%;
-  left: 15%;
+  top: 15%;
+  left: 12vw;
   display: grid;
   grid-template-columns: 5% repeat(3, 1fr) 5%;
   grid-template-rows: 50% 50%;
