+++ src/component/pages/DefaultLayout/DefaultLayout.module.css
@@ -1,13 +1,21 @@
 .container {
     display: grid;
-    grid-template-rows: 100%;
-    grid-template-columns: 10% 90%;
+    grid-template-rows: 92vh 8vh;
+    grid-template-columns: 10vw 1fr;
     position: relative;
     
 }
 
+.header {
+    grid-row-start: 1;
+    grid-column-start: 1;
+    grid-column-end: 3;
+    justify-self: center;
+}
+
+
 .background {
-    background-image: url(../../assets/detailedbackground.svg);
+    background-image: url(../../../assets/detailedbackground.svg);
 	background-size: cover;
 	background-position: center, center;
     transition: transform 1.2s;
@@ -17,16 +25,22 @@
     top: 5;
     left: -5%;
 }
-.list {
+
+.timeline {
+    grid-column-start: 1;
+    grid-row-start: 1;
+    position: relative;
+    align-self: center;
     display: grid;
     grid-template-rows: repeat(13, 1fr);
     justify-self: center;
     margin-top: 3.5vh;
-    background: rgba(255, 255, 255, 0.5);
+    background: rgba(216, 216, 216, 0.253);
     border-radius: 45px;
     width: 90px;
-    height: 90vh;
+    height: 80vh;
     text-align: center;
+    
 }
 
 .buttonArrowUp {
@@ -51,4 +65,9 @@
     border: none;
     background-color: transparent;
     outline: none;
+}
+
+.filter {
+    grid-column: span 2;
+    
 }
