+++ src/component/pages/Year1950/Year1950.js
@@ -1,11 +1,12 @@
 import React from "react";
-import style from "./Year1950.module.css"
-
-
+import style from "./Year1950.module.css";
+import InformationBoxLayout from "../../InformationBoxLayout/InformationBoxLayout";
 
 const Year1950 = () => {
+   
  return (
-    <div>
+    <div className={style.container}>
+         <InformationBoxLayout />
     </div>
  )
 }
