--- src/component/pages/Year1950/Year1950.module.css
@@ -1,1 +1,0 @@
-
