+++ src/component/pages/Year1955/Year1955.js
@@ -1,9 +1,10 @@
 import React from "react";
+import InformationBoxLayout from "../../InformationBoxLayout/InformationBoxLayout";
 
 const Year1955 = () => {
  return (
     <div>
-        <h2>1955</h2>
+        <InformationBoxLayout />
     </div>
  )
 }
