+++ src/component/pages/Year1956/Year1956.js
@@ -1,9 +1,11 @@
 import React from "react";
+import InformationBoxLayout from "../../InformationBoxLayout/InformationBoxLayout";
+
 
 const Year1956 = () => {
  return (
     <div>
-        <h2>1956</h2>
+        <InformationBoxLayout />
     </div>
  )
 }
