+++ src/component/pages/Year1958/Year1958.js
@@ -1,9 +1,11 @@
 import React from "react";
+import InformationBoxLayout from "../../InformationBoxLayout/InformationBoxLayout";
+
 
 const Year1958 = () => {
  return (
     <div>
-        <h2>1958</h2>
+        <InformationBoxLayout />
     </div>
  )
 }
