+++ src/component/pages/Year1959/Year1959.js
@@ -1,9 +1,11 @@
 import React from "react";
+import InformationBoxLayout from "../../InformationBoxLayout/InformationBoxLayout";
+
 
 const Year1959 = () => {
  return (
     <div>
-        <h2>1959</h2>
+        <InformationBoxLayout />
     </div>
  )
 }
