+++ src/component/pages/Year1960/Year1960.js
@@ -1,9 +1,10 @@
 import React from "react";
+import InformationBoxLayout from "../../InformationBoxLayout/InformationBoxLayout";
 
 const Year1960 = () => {
  return (
     <div>
-        <h2>1960</h2>
+        <InformationBoxLayout />
     </div>
  )
 }
