+++ src/timeline.json
@@ -1,0 +1,124 @@
+{
+	"timelineInfo": [
+		{
+			"year": "1950",
+			"category": "event",
+			"info": "Concept of resilience fi_x005F_x001f_rst noted in psychology, “Maternal care and mental health” by John Bowlby, WHO",
+			"link": "https://apps.who.int/iris/bitstream/handle/10665/40724/WHO_MONO_2_(part1).pdf;jsessionid=E24D41B690EEACD11A968BDB6359907F?sequence=1"
+		},
+		{
+			"year": "1973",
+			"category": "publication",
+			"info": "\"Competence and adaptation in adult schizophrenic patients and children at risk\" by Norman Garmezy",
+			"link": "https://www.annualreviews.org/doi/abs/10.1146/annurev.es.04.110173.000245"
+		},
+		{
+			"year": "1973",
+			"category": "publication",
+			"info": "“Resilience and Stability of Ecological Systems” by C.S. Holling",
+			"link": "https://www.annualreviews.org/doi/abs/10.1146/annurev.es.04.110173.000245"
+		},
+		{
+			"year": "1983",
+			"category": "trend",
+			"info": "Ethiopia famine and hunger crisis",
+			"link": ""
+		},
+		{
+			"year": "1985",
+			"category": "event",
+			"info": "Famine Early Warning Systems Network (FEWSNET) created",
+			"link": ""
+		},
+		{
+			"year": "1992",
+			"category": "trend",
+			"info": "UN Conference on Environment and Development (UNCED) in Rio, Brazil",
+			"link": ""
+		},
+		{
+			"year": "1995",
+			"category": "event",
+			"info": "UN Millennium Development Goals (MDGs)",
+			"link": ""
+		},
+		{
+			"year": "1995",
+			"category": "program, platform, investment initiative",
+			"info": "Sustainable Livelihoods Framework",
+			"link": "https://hhi.harvard.edu/education/atha/content/sustainable-livelihoods-framework"
+		},
+		{
+			"year": "1996",
+			"category": "trend",
+			"info": "Mobile phones and SMS enter the mainstream",
+			"link": ""
+		},
+		{
+			"year": "1999",
+			"category": "program, platform, investment initiative",
+			"info": "African Evaluation Association formed",
+			"link": "https://afrea.org/"
+		},
+		{
+			"year": "1999",
+			"category": "program, platform, investment initiative",
+			"info": "Resilience Alliance formed",
+			"link": "https://www.resalliance.org/background"
+		},
+		{
+			"year": "2001",
+			"category": "publication",
+			"info": "\"Panarchy: Understanding Transformations in Systems of Humans and Nature\"by Lance Gunderston & C.S. Holling",
+			"link": "www.resalliance.org/index.php/panarchy"
+		},
+		{
+			"year": "2004",
+			"category": "trend",
+			"info": "Facebook’s website launches",
+			"link": ""
+		},
+		{
+			"year": "2004",
+			"category": "trend",
+			"info": "ndian Ocean tsunami",
+			"link": ""
+		},
+		{
+			"year": "2005",
+			"category": "trend",
+			"info": "Kyoto Protocol enters into force, UN Framework Convention on Climate Change (UNFCCC) COP10",
+			"link": "https://en.wikipedia.org/wiki/2005_United_Nations_Climate_Change_Conference"
+		},
+		{
+			"year": "2005",
+			"category": "trend",
+			"info": "CBA1: First conference on community-based adaptation to climate change",
+			"link": "https://www.iied.org/1st-international-conference-community-based-adaptation-cba1"
+		},
+		{
+			"year": "2005",
+			"category": "trend",
+			"info": "Hurricane Katrina, a category 5 event, makes landfall on the Gulf Coast of the US",
+			"link": "https://www.rand.org/blog/2015/09/what-hurricane-katrina-taught-us-about-community-resilience.html"
+		},
+		{
+			"year": "2005",
+			"category": "event",
+			"info": "Government of Ethiopia launches the Productive Safety Net Programme (PSNP)",
+			"link": "https://www.wfp.org/sites/default/files/PSNP%20Factsheet.pdf"
+		},
+		{
+			"year": "2005",
+			"category": "framework: conceptual, evaluative, operational",
+			"info": "Hyogo Framework for Action led by UN Office for Disaster Risk Reduction (UNISDR)",
+			"link": "https://www.undrr.org/"
+		},
+		{
+			"year": "2006",
+			"category": "trend",
+			"info": "TREND: Influx of funds after the Indian Ocean tsunami, Katrina and Haiti, which spurs further work and dialogue around disaster resilience and allows more disaster risk reduction with a focus on resilience",
+			"link": ""
+		}
+	]
+}
